Add data_only mode, SD_ERR_STATE, and rename max_results to limit
- sd_stardict/sd_dictd: add data_only param to skip index loading
- Make open_from_paths static (file names are fixed)
- Add SD_ERR_STATE guard for lookup/suggest when index is NULL
- Change index_file_size/syn_file_size from uint64_t to uint32_t
- Rename max_results to limit across all suggest APIs

+++ src/sd_stardict_ifo.c
@@ -233,11 +233,11 @@ sd_stardict_ifo *sd_stardict_ifo_load(const char *ifo_filename) {
                 ifo->sametypesequence = value;
                 value = NULL;
             } else if (strcmp(key, "idxfilesize") == 0 && value) {
-                ifo->index_file_size = (uint64_t)atoll(value);
+                ifo->index_file_size = (uint32_t)atol(value);
             } else if (strcmp(key, "tdxfilesize") == 0 && value) {
-                ifo->index_file_size = (uint64_t)atoll(value);
+                ifo->index_file_size = (uint32_t)atol(value);
             } else if (strcmp(key, "synfilesize") == 0 && value) {
-                ifo->syn_file_size = (uint64_t)atoll(value);
+                ifo->syn_file_size = (uint32_t)atol(value);
             }
 
             free(key);
